distribute code duplication
